chore: split PaymentsController in 3 files
